+++ app/(main)/studio/edit/[[...id]]/layout.tsx
