+++ app/(main)/studio/edit/[[...id]]/page.tsx
