+++ app/(main)/studio/layout.tsx
@@ -1,0 +1,7 @@
+export default async function Layout({
+  children,
+}: {
+  children: React.ReactNode
+}) {
+  return <>{children}</>
+}
