+++ app/(main)/studio/loading.tsx
