+++ app/(main)/studio/page.tsx
