+++ app/(play)/play/[[...id]]/layout.tsx
