+++ app/(play)/play/[[...id]]/page.tsx
